+++ CHANGELOG.md
@@ -8,6 +8,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Added information about the Magento edition to the order request.
 
+### Fixed
+- Fixed error when creating transaction without available phone number
+
 ## [2.4.0] - 2021-03-11
 ### Added
 - Added 3 generic gateways and 3 generic giftcards.